Extracted a common Mixin and applied to subclasses
